+++ src/main/java/com/bitsu/social_media/service/ReactionService.java
@@ -6,6 +6,7 @@ import com.bitsu.social_media.dto.ReactionResponse;
 import com.bitsu.social_media.model.Reaction;
 import com.bitsu.social_media.repository.PostRepo;
 import com.bitsu.social_media.repository.ReactionRepo;
+import com.bitsu.social_media.utility.Utility;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -17,6 +18,7 @@ public class ReactionService {
     private final ReactionRepo reactionRepo;
     private final UserService userService;
     private final PostRepo postRepo;
+    private final Utility utility;
 
     public List<ReactionResponse> getReactions() {
         return reactionRepo.findAll().stream()
@@ -48,7 +50,7 @@ public class ReactionService {
 
         Reaction reaction = Reaction.builder()
                 .type(reactionRequest.getType())
-                .user(userService.getLoggedInUser())
+                .user(utility.getLoggedInUser())
                 .post(post)
                 .build();
