+++ src/main/java/com/bitsu/social_media/service/UserService.java
@@ -6,6 +6,7 @@ import com.bitsu.social_media.dto.UserProfilePic;
 import com.bitsu.social_media.dto.UserResponse;
 import com.bitsu.social_media.model.User;
 import com.bitsu.social_media.repository.UserRepo;
+import com.bitsu.social_media.utility.Utility;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
@@ -18,6 +19,7 @@ public class UserService {
 
     private final UserRepo userRepo;
     private final S3Service s3Service;
+    private final Utility utility;
 
     public UserResponse mapToUserResponse(User user) {
         return UserResponse.builder()
@@ -28,31 +30,21 @@ public class UserService {
                 .build();
     }
 
-    public User getLoggedInUser() {
-        return userRepo.findByUsername(
-                ((User) SecurityContextHolder
-                        .getContext()
-                        .getAuthentication()
-                        .getPrincipal()
-                ).getUsername()
-        ).orElseThrow(() -> new RuntimeException("User not found ** change for testing webhook **"));
-    }
-
     public void updatePI(UserPIInfo userPIInfo) {
-        User user = getLoggedInUser();
+        User user = utility.getLoggedInUser();
         user.setFirstname(userPIInfo.getFirstname());
         user.setLastname(userPIInfo.getLastname());
         userRepo.save(user);
     }
 
     public void updateBio(UserBioInfo userBioInfo) {
-        User user = getLoggedInUser();
+        User user = utility.getLoggedInUser();
         user.setBio(userBioInfo.getBio());
         userRepo.save(user);
     }
 
     public void updateProfilePic(UserProfilePic userProfilePic) {
-        User user = getLoggedInUser();
+        User user = utility.getLoggedInUser();
         s3Service.deleteImageFromBucket(userProfilePic.getProfilePicture());
         user.setProfilePicture(userProfilePic.getProfilePicture());
         userRepo.save(user);
@@ -71,25 +63,25 @@ public class UserService {
 
     public List<UserResponse> getFollowing(String search) {
         if (search == null || search.isBlank()) {
-            return getLoggedInUser().getFollowers().stream()
+            return utility.getLoggedInUser().getFollowers().stream()
                     .map(this::mapToUserResponse)
                     .toList();
         }
-        return getLoggedInUser().getFollowers().stream()
+        return utility.getLoggedInUser().getFollowers().stream()
                 .filter(user -> user.getUsername().contains(search))
                 .map(this::mapToUserResponse)
                 .toList();
     }
 
     public void unfollow(int id) {
-        User user = getLoggedInUser();
+        User user = utility.getLoggedInUser();
         User userToUnfollow = userRepo.findById(id).orElseThrow(() -> new RuntimeException("User to unfollow not found"));
         user.getFollowers().remove(userToUnfollow);
         userRepo.save(user);
     }
 
     public void follow(int id) {
-        User user = getLoggedInUser();
+        User user = utility.getLoggedInUser();
         User userToFollow = userRepo.findById(id).orElseThrow(() -> new RuntimeException("User to follow not found"));
         user.getFollowers().add(userToFollow);
         userRepo.save(user);
